backend/foodgram/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
psycopg2-binary==2.9.5
djoser==2.1.0
django-filter==21.1
django-colorfield==0.7.2
drf-extra-fields==3.4.0
drf-yasg==1.21.3
django-rest-swagger==2.2.0
gunicorn==20.0.4
python-dotenv==0.21.0
asgiref==3.3.2
drf-extra-fields==3.4.1
